requirements.txt
# HTML extraction (primary: trafilatura, fallback: readability + markdownify)
trafilatura>=2.0.0
markdownify>=1.2.2

# HTML content extraction (fallback tier)
readability-lxml>=0.8.4
lxml>=5.0.0

# Server framework
litserve>=0.2.16

# HTTP client (for client.py)
requests>=2.32.5

# Async HTTP client (for URL fetching)
httpx>=0.28.0

# Terminal formatting
rich>=14.3.0

# Environment configuration
python-dotenv>=1.2.0
